refactor (calculator): 테스트케이스 추가, main/테스트 패키지 추가
