+++ src/test/java/calculator/main/Parser.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.main;
 
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
