+++ src/test/java/calculator/main/StringCalculator.java
@@ -1,4 +1,4 @@
-package calculator;
+package calculator.main;
 
 import java.util.Arrays;
