+++ src/test/java/calculator/test/StringCalculatorTest.java
@@ -1,12 +1,17 @@
-package calculator;
+package calculator.test;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatExceptionOfType;
 
+import calculator.main.Parser;
+import calculator.main.StringCalculator;
+import java.util.stream.Stream;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
 import org.junit.jupiter.params.provider.NullAndEmptySource;
 import org.junit.jupiter.params.provider.ValueSource;
 
@@ -42,16 +47,33 @@ public class StringCalculatorTest {
 
     @DisplayName(value = "숫자 두개를 쉼표(,) 구분자로 입력할 경우 두 숫자의 합을 반환한다.")
     @ParameterizedTest
-    @ValueSource(strings = {"1,2"})
-    void twoNumbers(final String text) {
-        assertThat(calculator.add(text)).isSameAs(3);
+    @MethodSource("makeTextCommaDelimiter")
+    void twoNumbers(final String text, final int sum) {
+        assertThat(calculator.add(text)).isSameAs(sum);
+    }
+
+    static Stream<Arguments> makeTextCommaDelimiter() {
+        return Stream.of(
+            Arguments.of("1,2,3,4,5", 15),
+            Arguments.of("1,2,3,4,5,6", 21),
+            Arguments.of("1,2,3,4,5,6,7", 28)
+        );
     }
 
     @DisplayName(value = "구분자를 쉼표(,) 이외에 콜론(:)을 사용할 수 있다.")
     @ParameterizedTest
-    @ValueSource(strings = {"1,2:3,4,5"})
+    @MethodSource("makeTextAddColonDelimiter")
     void colons(final String text) {
         assertThat(calculator.add(text)).isSameAs(15);
+    }
+
+    static Stream<Arguments> makeTextAddColonDelimiter() {
+        return Stream.of(
+            Arguments.of("1:2,3,4,5"),
+            Arguments.of("1,2:3:4,5"),
+            Arguments.of("1,2,3:4:5"),
+            Arguments.of("1:2:3:4:5")
+        );
     }
 
     @DisplayName(value = "구분자를 쉼표(,)와 콜론(:) 이외에 값을 사용하면 RuntimeException 예외 처리를 한다.")
@@ -60,13 +82,6 @@ public class StringCalculatorTest {
     void invalid_delimiter(final String text) {
         assertThatExceptionOfType(RuntimeException.class)
             .isThrownBy(() -> calculator.add(text));
-    }
-
-    @DisplayName(value = "입력값이 모두 0인 경우, 합은 0이다.")
-    @ParameterizedTest
-    @ValueSource(strings = {"0:0:0,0,0:0"})
-    void input_multiple_zero(final String text) {
-        assertThat(calculator.add(text)).isSameAs(0);
     }
 
     @DisplayName(value = "구분자만 포함되고 값이 없는 경우, RuntimeException 예외 처리를 한다.")
@@ -95,9 +110,17 @@ public class StringCalculatorTest {
 
     @DisplayName(value = "//와 \n 문자 사이에 커스텀 구분자를 지정할 수 있다.")
     @ParameterizedTest
-    @ValueSource(strings = {"//-\n1-2-3-4-5-6-7-8-9-10"})
+    @MethodSource("makeTextIncludeCustomDelimiter")
     void customDelimiter(final String text) {
         assertThat(calculator.add(text)).isSameAs(55);
+    }
+
+    static Stream<Arguments> makeTextIncludeCustomDelimiter() {
+        return Stream.of(
+            Arguments.of("//-\n1-2-3-4-5-6-7-8-9-10"),
+            Arguments.of("//+\n1+2+3+4+5+6+7+8+9+10"),
+            Arguments.of("//*\n1*2*3*4*5*6*7*8*9*10")
+        );
     }
 
     @DisplayName(value = "커스텀 구분자는 한개를 지정할 수 있으며, 복수개를 지정하는 경우 RuntimeException 예외 처리를 한다.")
@@ -106,12 +129,6 @@ public class StringCalculatorTest {
     void multiple_custom_delimiter(final String text) {
         assertThatExceptionOfType(RuntimeException.class)
             .isThrownBy(() -> calculator.add(text));
-    }
-
-    @DisplayName(value = "커스텀 구분자를 사용하면서, 입력이 0인 경우 합은 0이다.")
-    @Test
-    void zero_custom_delimiter() {
-        assertThat(calculator.add("//*\n0*0*0*0*0")).isZero();
     }
 
     @DisplayName(value = "커스텀 구분자를 사용하면서, 음수를 전달하는 경우 RuntimeException 예외 처리를 한다.")
